Public folder, object oriented config, mime, naming
Static file server mapping public folder, mime type with node-mime, html
file class, id naming, button click events removed with form submit
events for login and message forms. Exit alert box shows only logged in
users. README.
