+++ README.md
@@ -1,11 +1,9 @@
 #NodHat
 
 Simple Nodejs Chat Application. 
-Demo available on http://chat.vefasizalem.org/
 
 TODO:
-- Recent messages store on Redis or MongoDB.
-- Cross browser compability.
-- Smiley, image, url parsing.
-- Txt validation, sanitization.
-- Clean user session.
+- Notification sound on/off option
+- Smiley and image importing
+- Simple UI
+
