--- assets/css/style.css
@@ -1,189 +1,0 @@
-* {
-    margin: 0;
-    padding: 0;
-    outline: none;
-}
-a {
-    text-decoration: none;
-}
-body {
-    overflow: hidden;
-    font-family: Arial, Helvetica, sans-serif;
-    background: #1D1F21;
-}
-#main {
-    position: absolute;
-    top: 0;
-    bottom: 0px;
-    left: 0px;
-    height: 100%;
-    width: 80%;
-}
-#users,
-#main {
-    display: none;
-}
-.clearfix {
-    clear: both;
-}
-#giris {
-    top: 50%;
-    left: 50%;
-    position: absolute;
-    width: 300px;
-    margin: -56px 0 0 -171px;
-}
-#giris h1 {
-    color: #ff7900;
-}
-#giris h1.logo {
-    background: url("../img/logo.png") no-repeat;
-    text-indent: -9999px;
-}
-#kul {
-    margin-right: 20px;
-    font-size: 13px;
-    font-weight: bold;
-    color: #474646;
-}
-#username {
-    float: left;
-    padding: 5px;
-    width: 200px;
-    border: none;
-    height: 22px;
-}
-#loginButton {
-    float: left;
-    height: 32px;
-    width: 90px;
-    border: none;
-    background: #78a300;
-}
-#users {
-    z-index: 2;
-    position: absolute;
-    right: 0;
-    height: 100%;
-    width: 20%;
-    border-width: 0 0 1px 1px;
-    background: #47535b;
-}
-#users h3 {
-    background: #9dbc7a;
-    padding: 5px;
-    color: #444;
-    font-size: 14px;
-}
-#users p {
-    padding: 2px 5px;
-    font-weight: bold;
-    font-size: 12px;
-}
-#users_list {
-    list-style-type: none;
-    padding: 5px;
-    overflow: auto;
-    /* position: absolute; */
-    height: 80%;
-}
-#users_list li {
-    background: #434F57;
-    padding: 2px;
-    margin-top: 1px;
-}
-#users_list li span {
-    color: #ffcc00;
-    font-size: .8em;
-}
-#users_list li i {
-    color: #52b043
-}
-#mesajlar {
-    overflow: auto;
-    height: 80%;
-}
-#mesaj {
-    padding: 10px;
-}
-#mesaj li {
-    list-style-type: none;
-    font-size: 14px;
-    padding: 3px;
-    color: #eeebac;
-}
-.message {
-    color: #c5c8c6;
-    -ms-word-break: break-all;
-    word-break: break-word;
-    word-wrap: break-word;
-    -webkit-hyphens: auto;
-    -moz-hyphens: auto;
-    hyphens: auto;
-}
-.message a {
-    color: #e53b2c
-}
-.username {
-    font-weight: bold;
-    color: #c68143;
-    font-size: 0.9em;
-}
-li.system .username {
-    color: rgb(170, 0, 0);
-}
-.status {
-    right: 0;
-    width: 15px;
-    height: 100%;
-    position: absolute;
-    background: #9fbb58;
-    /*#e64b50; */
-    top: 0;
-}
-li.system .message {
-    font-size: 0.9em;
-    color: #999;
-    font-style: italic;
-}
-#gonderici {
-    height: 20%;
-    border-width: 1px 0 0 0;
-    background: #444;
-    position: relative;
-}
-#gonderic {
-    padding: 10px 10px 30px 10px;
-    overflow: auto
-}
-#icerik {
-    float: left;
-    padding: 0 10px;
-    width: 50%;
-    border: 0;
-    background: #e1e8ed;
-    height: 29px;
-}
-#gonder {
-    float: left;
-    margin: 0 10px;
-    height: 29px;
-    width: 100px;
-    background: #82b541;
-    border: 0;
-    color: #333;
-}
-#copy {
-    position: absolute;
-    bottom: 0;
-    font-weight: normal;
-}
-#copy a {
-    color: #000;
-}
-#uyari {
-    line-height: 24px;
-    font-size: 12px;
-    font-weight: bold;
-    color: #dc5034;
-}
