--- index.html
@@ -1,58 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <title>Nodhat</title>
-    <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
-    <link type="text/css" rel="stylesheet" href="http://necolas.github.io/normalize.css/3.0.1/normalize.css">
-    <link type="text/css" rel="stylesheet" href="assets/css/style.css" />
-    <link type="text/css" rel="stylesheet" href="http://fortawesome.github.io/Font-Awesome/assets/font-awesome/css/font-awesome.css">
-    <!--[if IE 7]><link rel="stylesheet" href='http://fontawesome.io/3.2.1/assets/font-awesome/css/font-awesome-ie7.min.css'/><![endif]-->
-</head>
-
-<body>
-    <div id="giris" class="grad">
-        <h1 class="logo">NodHat</h1>
-        <input id="username" type="text" placeholder="nick" />
-        <button id="loginButton" onclick="javascript:void(0);"><i class="fa fa-send logo"></i>
-        </button>
-        <div id="uyari"></div>
-        <div class="clearfix"></div>
-    </div>
-    <section id="users">
-        <h3><i class="fa fa-users"></i> <span class="users_count"></span> &middot; Online </h3>
-        <ul id="users_list"></ul>
-        <div class="copy">
-            <p id="copy">Available on <a href="https://github.com/enverarslan/Nodhat" target="_blank">Github</a>.
-                <br/>Developed by <a href="http://vefasizalem.com" target="blank">Enver Arslan</a> with <a href="http://nodejs.org" target="blank">Node.JS</a>
-        </div>
-        </p>
-    </section>
-    <section id="main">
-        <div id="mesajlar">
-            <ul id="mesaj"></ul>
-        </div>
-        <div id="gonderici">
-            <div id="gonderic">
-                <input type="text" id="icerik" />
-                <button id="gonder"><i class="fa fa-send"></i> Send</button>
-                <div class="status" title="Connected, lets talk!"></div>
-                <div class="clearfix"></div>
-            </div>
-
-        </div>
-    </section>
-    <audio id="audio">
-        <source src="assets/sounds/chat.mp3" type="audio/mpeg" />
-        <source src="assets/sounds/chat.wav" type="audio/wav" />
-    </audio>
-
-    <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js" type="text/javascript"></script>
-    <script src="assets/js/string.js" type="text/javascript"></script>
-    <script src="assets/js/animate-colors-min.js" type="text/javascript"></script>
-    <script src="assets/js/urlize.js" type="text/javascript"></script>
-    <script src="http://localhost:1200/socket.io/socket.io.js"></script>
-    <script src="assets/js/client.js" type="text/javascript"></script>
-</body>
-
-</html>
