+++ lib/db.js
@@ -1,5 +1,5 @@
 var mongoose = require('mongoose'),
-    config = require('../config/config')(),
+    config = require('../config/host'),
     Schemas = require('../config/db/schemas.js');
 
 mongoose.connect(config.connectionString());
